Split the pointcloud filters into one module per responsibility
filters.py had grown to hold two unrelated transforms plus the helpers they
happened to share, which made it the default landing place for anything
sensing-related. Split along responsibility lines:

  nebula_mask.py    NebulaDownsampleMaskFilter and its mask/calibration handling
  crop_box.py       CropBoxFilter and load_crop_boxes
  lidar_sources.py  splitting a concatenated cloud back into per-LiDAR slices
  assets.py         locating bundled assets

ego_motion.py is unchanged. Tests follow the same split: test_nebula_mask.py,
test_crop_box.py and test_ego_motion.py, leaving test_point_cloud.py to the
transforms it actually covers.

The shared helpers were private to filters.py and are now part of a module
interface, so they lose their underscore and gained docstrings:
_SourceSlice -> SourceSlice, _iter_pointcloud_sources -> iter_pointcloud_sources,
_iter_concat_sources -> iter_concat_sources, _normalize_lidar_name ->
normalize_lidar_name, _resolve_path -> assets.resolve_asset_path. The eight
mounting-position names are now LIDAR_POSITION_NAMES in lidar_sources rather than
being read back out of the mask-path dict.

Pure refactor: no behaviour change, and the end-to-end reconstruction still
gives +0.318% point count and 0.8771 voxel IoU.
