+++ autoware_ml/tests/transforms/test_point_cloud.py
@@ -1,5 +1,3 @@
-import json
-
 import numpy as np
 import pytest
 
@@ -9,18 +7,6 @@ from autoware_ml.transforms.point_cloud.crop import (
     CropBoxOuter,
     PointsRangeFilter,
     SphereCrop,
-)
-from autoware_ml.transforms.point_cloud.ego_motion import (
-    PRE_CORRECTION_POINTS_KEY,
-    InvertEgoMotionCorrection,
-    PoseTable,
-    invert_ego_motion,
-    normalize_quaternions,
-)
-from autoware_ml.transforms.point_cloud.filters import (
-    CropBoxFilter,
-    NebulaDownsampleMaskFilter,
-    load_crop_boxes,
 )
 from autoware_ml.transforms.point_cloud.geometry import (
     GlobalRotScaleTrans,
@@ -36,7 +22,6 @@ from autoware_ml.transforms.point_cloud.sampling import (
     RandomDropout,
 )
 from autoware_ml.transforms.point_cloud.sweeps import LoadPointsFromMultiSweeps
-
 
 _TEST_CROP_BOXES = [
     [-1.5, -1.2, 0.0, 5.7, 1.2, 3.0],  # body-like
@@ -223,326 +208,6 @@ class TestPointCloudTransforms:
 
         assert output["coord"].shape[0] == 4
         assert output["strength"].shape[0] == 4
-
-    def test_nebula_downsample_mask_filter_keeps_aligned_point_arrays(self, tmp_path):
-        mask_root = tmp_path / "masks"
-        calibration_root = tmp_path / "calibrations"
-        (mask_root / "front_upper").mkdir(parents=True)
-        calibration_root.mkdir()
-        mask = np.zeros((2, 10), dtype=np.uint8)
-        mask[0, :] = 255
-        cv2 = pytest.importorskip("cv2")
-        cv2.imwrite(str(mask_root / "front_upper" / "mask.png"), mask)
-        (calibration_root / "model.csv").write_text(
-            "Channel,Elevation,Azimuth\n1,0.0,0.0\n2,45.0,0.0\n"
-        )
-
-        sample = {
-            "points": np.array(
-                [
-                    [10.0, 0.0, 0.0, 1.0],
-                    [10.0, 0.0, 10.0, 2.0],
-                ],
-                dtype=np.float32,
-            ),
-            "pts_semantic_mask": np.array([7, 9], dtype=np.int64),
-            "source_name": "LIDAR_FRONT_UPPER",
-        }
-
-        output = NebulaDownsampleMaskFilter(
-            mask_root=str(mask_root),
-            calibration_root=str(calibration_root),
-            lidar_name_to_mask={"front_upper": "front_upper/mask.png"},
-            lidar_name_to_model={"front_upper": "test_model"},
-            lidar_name_to_calibration={"front_upper": "model.csv"},
-            use_calibration_azimuth_offsets=False,
-            return_stats=True,
-        )(sample)
-
-        assert output["points"].shape == (1, 4)
-        np.testing.assert_array_equal(output["pts_semantic_mask"], np.array([7], dtype=np.int64))
-        assert output["nebula_downsample_stats"] == [
-            {"source_name": "LIDAR_FRONT_UPPER", "num_input_points": 2, "num_kept_points": 1}
-        ]
-
-    def test_nebula_downsample_mask_filter_uses_concat_source_slices(self, tmp_path):
-        mask_root = tmp_path / "masks"
-        calibration_root = tmp_path / "calibrations"
-        for name, value in {"front_upper": 255, "rear_upper": 0}.items():
-            (mask_root / name).mkdir(parents=True)
-            cv2 = pytest.importorskip("cv2")
-            cv2.imwrite(str(mask_root / name / "mask.png"), np.full((1, 10), value, np.uint8))
-        calibration_root.mkdir()
-        (calibration_root / "model.csv").write_text("Channel,Elevation,Azimuth\n1,0.0,0.0\n")
-
-        sample = {
-            "points": np.array([[10.0, 0.0, 0.0, 1.0], [20.0, 0.0, 0.0, 2.0]], dtype=np.float32),
-            "labels": np.array([1, 2], dtype=np.int64),
-            "lidar_sources": {
-                "LIDAR_FRONT_UPPER": {
-                    "sensor_token": "front",
-                    "translation": [0.0, 0.0, 0.0],
-                    "rotation": np.eye(3).tolist(),
-                },
-                "LIDAR_REAR_UPPER": {
-                    "sensor_token": "rear",
-                    "translation": [0.0, 0.0, 0.0],
-                    "rotation": np.eye(3).tolist(),
-                },
-            },
-            "lidar_sources_info": {
-                "sources": [
-                    {"sensor_token": "front", "idx_begin": 0, "length": 1},
-                    {"sensor_token": "rear", "idx_begin": 1, "length": 1},
-                ]
-            },
-        }
-
-        output = NebulaDownsampleMaskFilter(
-            mask_root=str(mask_root),
-            calibration_root=str(calibration_root),
-            lidar_name_to_mask={
-                "front_upper": "front_upper/mask.png",
-                "rear_upper": "rear_upper/mask.png",
-            },
-            lidar_name_to_model={"front_upper": "test_model", "rear_upper": "test_model"},
-            lidar_name_to_calibration={"front_upper": "model.csv", "rear_upper": "model.csv"},
-            use_calibration_azimuth_offsets=False,
-        )(sample)
-
-        np.testing.assert_allclose(output["points"], np.array([[10.0, 0.0, 0.0, 1.0]]))
-        np.testing.assert_array_equal(output["labels"], np.array([1], dtype=np.int64))
-
-    def test_nebula_downsample_mask_filter_uses_nebula_azimuth_convention(self, tmp_path):
-        mask_root = tmp_path / "masks"
-        calibration_root = tmp_path / "calibrations"
-        (mask_root / "front_upper").mkdir(parents=True)
-        calibration_root.mkdir()
-        cv2 = pytest.importorskip("cv2")
-        mask = np.zeros((1, 4), dtype=np.uint8)
-        mask[0, 0] = 255
-        cv2.imwrite(str(mask_root / "front_upper" / "mask.png"), mask)
-        (calibration_root / "model.csv").write_text("Channel,Elevation,Azimuth\n1,0.0,0.0\n")
-
-        sample = {
-            "points": np.array(
-                [
-                    [0.0, 10.0, 0.0, 1.0],
-                    [10.0, 0.0, 0.0, 2.0],
-                ],
-                dtype=np.float32,
-            ),
-            "source_name": "LIDAR_FRONT_UPPER",
-        }
-
-        output = NebulaDownsampleMaskFilter(
-            mask_root=str(mask_root),
-            calibration_root=str(calibration_root),
-            lidar_name_to_mask={"front_upper": "front_upper/mask.png"},
-            lidar_name_to_model={"front_upper": "test_model"},
-            lidar_name_to_calibration={"front_upper": "model.csv"},
-            use_calibration_azimuth_offsets=False,
-        )(sample)
-
-        np.testing.assert_allclose(output["points"], np.array([[0.0, 10.0, 0.0, 1.0]]))
-
-    def test_nebula_downsample_mask_filter_prefers_the_stored_ring_index(self, tmp_path):
-        mask_root = tmp_path / "masks"
-        calibration_root = tmp_path / "calibrations"
-        (mask_root / "front_upper").mkdir(parents=True)
-        calibration_root.mkdir()
-        cv2 = pytest.importorskip("cv2")
-        # Only ring 1 is kept. Both points sit at elevation 0, so estimating the ring from
-        # elevation would put both on ring 0 and drop them; the stored index must win.
-        mask = np.zeros((2, 4), dtype=np.uint8)
-        mask[1, 0] = 255
-        cv2.imwrite(str(mask_root / "front_upper" / "mask.png"), mask)
-        (calibration_root / "model.csv").write_text(
-            "Channel,Elevation,Azimuth\n1,0.0,0.0\n2,45.0,0.0\n"
-        )
-
-        # Columns are [x, y, z, intensity, channel].
-        sample = {
-            "points": np.array(
-                [[0.0, 10.0, 0.0, 1.0, 1.0], [0.0, 10.0, 0.0, 2.0, 0.0]],
-                dtype=np.float32,
-            ),
-            "source_name": "LIDAR_FRONT_UPPER",
-        }
-
-        kwargs = {
-            "mask_root": str(mask_root),
-            "calibration_root": str(calibration_root),
-            "lidar_name_to_mask": {"front_upper": "front_upper/mask.png"},
-            "lidar_name_to_model": {"front_upper": "test_model"},
-            "lidar_name_to_calibration": {"front_upper": "model.csv"},
-        }
-        output = NebulaDownsampleMaskFilter(channel_dim=4, **kwargs)(dict(sample))
-        np.testing.assert_allclose(output["points"], sample["points"][:1])
-
-        # Opting out falls back to the elevation estimate, which drops both points.
-        fallback = NebulaDownsampleMaskFilter(channel_dim=None, **kwargs)(dict(sample))
-        assert fallback["points"].shape[0] == 0
-
-    def test_nebula_downsample_mask_filter_does_not_apply_azimuth_offsets_by_default(self):
-        # The driver indexes the mask by raw azimuth; subtracting the per-channel calibration
-        # offsets shifts points into the wrong column.
-        assert (
-            NebulaDownsampleMaskFilter(mask_root="aip_x2_gen2").use_calibration_azimuth_offsets
-            is False
-        )
-
-    def test_crop_box_filter_removes_points_inside_any_box_and_keeps_arrays_aligned(self):
-        sample = {
-            "points": np.array(
-                [
-                    [0.0, 0.0, 0.5, 1.0, 0.0],  # inside the vehicle body
-                    [4.7, 0.0, 0.4, 2.0, 0.0],  # inside the wheels box
-                    [10.0, 0.0, 0.5, 3.0, 0.0],  # ahead of the vehicle
-                    [0.0, 0.0, -0.1, 4.0, 0.0],  # below the box floor (ground return)
-                ],
-                dtype=np.float32,
-            ),
-            "labels": np.array([10, 11, 12, 13], dtype=np.int64),
-        }
-
-        output = CropBoxFilter(crop_boxes=_TEST_CROP_BOXES)(sample)
-
-        np.testing.assert_allclose(
-            output["points"],
-            np.array(
-                [[10.0, 0.0, 0.5, 3.0, 0.0], [0.0, 0.0, -0.1, 4.0, 0.0]],
-                dtype=np.float32,
-            ),
-        )
-        np.testing.assert_array_equal(output["labels"], np.array([12, 13], dtype=np.int64))
-
-    def test_crop_box_filter_decides_on_pre_correction_points(self):
-        # The point sits outside the boxes where T4Dataset put it, but inside them where it
-        # actually was when captured. The vehicle crops pre-correction, so it must be dropped --
-        # and the surviving point must keep its corrected coordinates, not the pre-correction ones.
-        sample = {
-            "points": np.array(
-                [[0.0, 0.0, 0.5, 1.0, 0.0], [30.0, 0.0, 0.5, 2.0, 0.0]], dtype=np.float32
-            ),
-            PRE_CORRECTION_POINTS_KEY: np.array(
-                [[30.0, 0.0, 0.5], [1.0, 0.0, 0.5]], dtype=np.float32
-            ),
-        }
-
-        output = CropBoxFilter(crop_boxes=_TEST_CROP_BOXES)(sample)
-
-        np.testing.assert_allclose(
-            output["points"], np.array([[0.0, 0.0, 0.5, 1.0, 0.0]], dtype=np.float32)
-        )
-        # The parallel array is filtered by the same mask, so the two never drift apart.
-        assert output[PRE_CORRECTION_POINTS_KEY].shape[0] == output["points"].shape[0]
-        np.testing.assert_allclose(
-            output[PRE_CORRECTION_POINTS_KEY], np.array([[30.0, 0.0, 0.5]], dtype=np.float32)
-        )
-
-    def test_filters_reject_misaligned_pre_correction_points(self):
-        sample = {
-            "points": np.zeros((3, 5), dtype=np.float32),
-            PRE_CORRECTION_POINTS_KEY: np.zeros((2, 3), dtype=np.float32),
-        }
-
-        with pytest.raises(ValueError, match="aligned"):
-            CropBoxFilter(crop_boxes=_TEST_CROP_BOXES)(sample)
-
-    def test_invert_ego_motion_correction_round_trips_a_stationary_vehicle(self):
-        # With the ego pose constant across the sweep, undoing the correction is the identity.
-        poses = [
-            {
-                "timestamp": int((10.0 + i * 0.01) * 1e6),
-                "translation": [5.0, -2.0, 0.0],
-                "rotation": [1.0, 0.0, 0.0, 0.0],
-            }
-            for i in range(20)
-        ]
-        table = PoseTable(
-            times=np.array([p["timestamp"] * 1e-6 for p in poses], dtype=np.float64),
-            translations=np.array([p["translation"] for p in poses], dtype=np.float64),
-            quaternions=normalize_quaternions(
-                np.array([p["rotation"] for p in poses], dtype=np.float64)
-            ),
-        )
-        points = np.array([[1.0, 2.0, 3.0], [-4.0, 5.0, 6.0]], dtype=np.float32)
-        times = np.array([10.05, 10.09], dtype=np.float64)
-
-        out = invert_ego_motion(points, 10.02, times, table)
-
-        np.testing.assert_allclose(out, points, atol=1e-6)
-
-    def test_invert_ego_motion_correction_attaches_aligned_points(self, tmp_path):
-        scene = tmp_path / "scene"
-        (scene / "annotation").mkdir(parents=True)
-        (scene / "data").mkdir()
-        # Vehicle drives +x at 10 m/s; the sweep reference is t=10.0.
-        (scene / "annotation" / "ego_pose.json").write_text(
-            json.dumps(
-                [
-                    {
-                        "timestamp": int((10.0 + i * 0.01) * 1e6),
-                        "translation": [10.0 * (i * 0.01), 0.0, 0.0],
-                        "rotation": [1.0, 0.0, 0.0, 0.0],
-                    }
-                    for i in range(20)
-                ]
-            )
-        )
-        # Columns [x, y, z, intensity, channel, return_type, timestamp_ns].
-        points = np.zeros((2, 7), dtype=np.float32)
-        points[:, 0] = [20.0, 20.0]
-        points[:, 6] = [0.0, 50_000_000.0]  # 0 ms and 50 ms into the sweep
-        sample = {
-            "points": points,
-            "lidar_path": str(scene / "data" / "00000.pcd.bin"),
-            "lidar_sources_info": {
-                "stamp": {"sec": 10, "nanosec": 0},
-                "sources": [{"idx_begin": 0, "length": 2, "stamp": {"sec": 10, "nanosec": 0}}],
-            },
-        }
-
-        output = InvertEgoMotionCorrection(timestamp_dim=6)(sample)
-
-        pre = output[PRE_CORRECTION_POINTS_KEY]
-        assert pre.shape == (2, 3)
-        # The first point was captured at the reference time, so it does not move. The second was
-        # captured 50 ms later, by which time the vehicle had advanced 0.5 m, so relative to the
-        # ego frame at that instant the point sits 0.5 m nearer.
-        np.testing.assert_allclose(pre[0], [20.0, 0.0, 0.0], atol=1e-4)
-        np.testing.assert_allclose(pre[1], [19.5, 0.0, 0.0], atol=1e-4)
-
-    def test_invert_ego_motion_correction_requires_metadata_when_strict(self):
-        sample = {"points": np.zeros((2, 7), dtype=np.float32)}
-
-        with pytest.raises(KeyError):
-            InvertEgoMotionCorrection()(sample)
-
-        passthrough = InvertEgoMotionCorrection(strict=False)(dict(sample))
-        assert PRE_CORRECTION_POINTS_KEY not in passthrough
-
-    def test_crop_box_filter_rejects_malformed_boxes(self):
-        with pytest.raises(ValueError, match="at least one box"):
-            CropBoxFilter(crop_boxes=[])
-        with pytest.raises(ValueError, match="min <= max"):
-            CropBoxFilter(crop_boxes=[[1.0, 0.0, 0.0, -1.0, 1.0, 1.0]])
-
-    def test_bundled_crop_boxes_asset_loads(self):
-        # Pins the values in configs/assets/aip_x2_gen2/crop_boxes.param.yaml, whose header records
-        # how they were derived from that platform's vehicle dimensions.
-        boxes = load_crop_boxes("aip_x2_gen2/crop_boxes.param.yaml")
-
-        assert len(boxes) == 2
-        np.testing.assert_allclose(
-            boxes[0], [-1.525790, -1.226830, 0.0, 5.711110, 1.200580, 3.080], atol=1e-6
-        )
-        np.testing.assert_allclose(
-            boxes[1], [4.372418, -1.225794, 0.0, 5.147822, 1.225794, 0.8195], atol=1e-6
-        )
-        # Usable directly as the transform's only configuration.
-        assert CropBoxFilter(crop_boxes=boxes).crop_boxes.shape == (2, 6)
 
     def test_grid_sample_keeps_arrays_aligned(self):
         sample = {
